Rewrite Parameters into Params class
This completely rewrites and test-drives the Params class, now it only
handles parameters, and matching them with arguments is done in Fn.
Params has the responsibility of validating the parameter list
(raising all of the errors that would occur at function definition
time), and leaves raising all of the runtime arity errors to the Fn
class. Although Fn delegates to Params to find out if a number of
arguments matches the arity, Params only returns a boolean and passes on
what to do with this information.

This gets rid of the wonky inheritance implementation, perhaps
inheritance is still a good answer to this situation but I think the
solution I came up with before was too ugly to keep, I find this one
nicer even though it does use a few if-statements, if nothing less it is
a better starting point for refactoring into using inheritance better in
the future.