Require static path by default, opt in to relative
Paths in require are now relative to where `lasp` was called by default,
unless a second truthy argument is passed to `require`, in which case
they are relative to the current file.

This also makes a `__FILE__` variable available whenever executing a
file containing the path of the current file.