Use stdlib by default
Previously you had to explicitly pass Lasp.env_with_stdlib to have
access to it, but this is almost always what you want so this is now the
default.

One spec had to be refactored to make this work. I also extracted a
do_execute method to clean up the execute_file method a bit.